feat(rust): give the Rust crate the RangeType enum the other three ports have
`CandleSetting.range_type` was a bare `i32` carrying 0/1/2 in a doc comment,
while C has `TA_RangeType_*`, and Java and C# both have a `RangeType` enum on
the same field. Rust was the only port where the caller wrote a magic number
and the library checked it at run time.

It is now `RangeType::{RealBody, HighLow, Shadows}`, so the domain is enforced
by the type. The builder's `0..=2` check is gone -- there is no longer a value
for it to reject -- and the rejection moves to `RangeType::try_from(i32)`, the
one boundary a raw integer still crosses: the JSON-RPC server, a config file,
C interop. The server converts there and answers the same "Invalid candle
setting" it did before.

Generated candle unpacking casts to the i32 the range comparisons run on
(`.range_type as i32`), matching what Java's `.ordinal()` and C#'s `(int)` have
always done -- so only the Rust backend's output moves: 57 CDL files, types.rs
and the Rust server, with C, Java and C# byte-identical.

The unit test that fed the builder an out-of-domain range type is replaced by
one over the conversion, since the old input is now unconstructible. Sabotage-
proved rather than assumed: adding `3 => Self::Shadows` to the template,
regenerating, and confirming the arm reached the generated crate makes it fail
on `Ok(Shadows) != Err(BadParam)`.

Gates: cargo test --lib (49), --doc (350), clippy --all-targets and cargo doc
clean, generator suite (all green), ta_regtest --codegen c+rust 161/161 each
including the candle-settings matrix (5 rows x 61 functions, 108 settings
pushed over RPC, 196 outputs moved), and a second full generate is a no-op.

+++ ta_codegen/output/rust/library/src/ta_func/cdl2crows.rs
@@ -68,7 +68,7 @@ impl Core {
     /// the first output value can be produced.
     pub fn CDL2CROWS_Lookback(&self) -> usize {
         #[allow(non_snake_case)]
-        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type;
+        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type as i32;
         #[allow(non_snake_case)]
         let BodyLong_avgPeriod: i32 = self.candle_settings.body_long.avg_period;
         #[allow(non_snake_case)]
@@ -177,7 +177,7 @@ impl Core {
         let mut BodyLongTrailingIdx: usize = 0_usize;
         let mut lookbackTotal: usize = 0_usize;
         #[allow(non_snake_case)]
-        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type;
+        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type as i32;
         #[allow(non_snake_case)]
         let BodyLong_avgPeriod: i32 = self.candle_settings.body_long.avg_period;
         #[allow(non_snake_case)]
@@ -367,7 +367,7 @@ impl CDL2CROWS_StreamState {
 impl Core {
     fn CDL2CROWS_step_internal(&self, sp: &mut CDL2CROWS_StreamState, inOpen: f64, inHigh: f64, inLow: f64, inClose: f64, outInteger: &mut i32) {
         #[allow(non_snake_case)]
-        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type;
+        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type as i32;
         #[allow(non_snake_case)]
         let BodyLong_avgPeriod: i32 = self.candle_settings.body_long.avg_period;
         #[allow(non_snake_case)]
@@ -465,7 +465,7 @@ impl Core {
         let mut BodyLongTrailingIdx: usize = 0_usize;
         let mut lookbackTotal: usize = 0_usize;
         #[allow(non_snake_case)]
-        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type;
+        let BodyLong_rangeType: i32 = self.candle_settings.body_long.range_type as i32;
         #[allow(non_snake_case)]
         let BodyLong_avgPeriod: i32 = self.candle_settings.body_long.avg_period;
         #[allow(non_snake_case)]
